+++ packages/component/src/ScrollToBottom/InternalContext.js
@@ -3,7 +3,8 @@ import React from 'react';
 const context = React.createContext({
   offsetHeight: 0,
   scrollHeight: 0,
-  setTarget: () => 0
+  setTarget: () => 0,
+  styleToClassName: () => ''
 });
 
 context.displayName = 'ScrollToBottomInternalContext';