+++ packages/component/src/hooks/internal/useFunctionContext.js
@@ -1,0 +1,7 @@
+import { useContext } from 'react';
+
+import FunctionContext from '../../ScrollToBottom/FunctionContext';
+
+export default function useFunctionContext() {
+  return useContext(FunctionContext);
+}