+++ packages/component/src/hooks/internal/useStyleToClassName.js
@@ -1,0 +1,7 @@
+import useInternalContext from './useInternalContext';
+
+export default function useStyleToClassName() {
+  const { styleToClassName } = useInternalContext();
+
+  return styleToClassName;
+}