+++ packages/playground/src/index.js
@@ -3,5 +3,6 @@ import ReactDOM from 'react-dom';
 import App from './App';
 import registerServiceWorker from './registerServiceWorker';
 
-ReactDOM.render(<App />, document.getElementById('root'));
+ReactDOM.render(<App nonce="a1b2c3d" />, document.getElementById('root'));
+
 registerServiceWorker();